chore: refactor chemical store
