+++ src/components/Common/ChemicalTable/ViewLogModal.tsx
@@ -7,9 +7,10 @@ import { X } from "react-feather";
 const ViewLogsModalComp = () => {
   const store = useStore();
   const chemicalStore = store?.chemicals;
-  const showModal = chemicalStore?.showViewChemicalLogModal;
-  const onClose = () => (chemicalStore.showViewChemicalLogModal = false);
-  const chemical = chemicalStore.selectedChemical;
+  const chemicalModel = chemicalStore?.chemicalModel;
+  const showModal = chemicalModel?.showViewChemicalLogModal;
+  const onClose = () => (chemicalModel.showViewChemicalLogModal = false);
+  const chemical = chemicalModel.selectedChemical;
   return (
     showModal && (
       <div className="fixed inset-0 flex items-center justify-center z-50">
