+++ src/store/models/index.ts
@@ -1,1 +1,2 @@
 export * from "./labModel";
+export * from "./chemicalModel";
